# api/openapi.yaml
openapi: 3.0.3
info:
  title: GophKeeper API
  description: API взаимодействия CLI-клиента и Сервера GophKeeper.
  version: 1.0.0
servers:
  - url: https://localhost:8080
    description: Локальное окружение
...
paths:
...
  /api/v1/auth/register:
    post:
      summary: Зарегистрировать пользователя
      description: >-
        Создаёт нового пользователя. Перед проверкой login очищается от пробельных символов по краям,
        ASCII-буквы A-Z переводятся в нижний регистр. После канонизации login должен содержать от 3 до 32
        ASCII-символов. Password должен содержать от 8 до 64 печатных ASCII-символов от ! до ~ без пробелов.
        Password не возвращается в ответе. Максимальный размер тела запроса — 4 MiB.
      operationId: registerUser
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              login: alice
              password: correct-horse-battery-staple
      responses:
        '201':
          description: Пользователь зарегистрирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              example:
                id: 42
                login: alice
                created_at: '2026-07-01T12:00:00Z'
        '400':
          description: Некорректный JSON, login или password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidRequest:
                  summary: Некорректные регистрационные данные
                  value:
                    code: invalid_request
                    message: invalid registration data
        '409':
          description: Login уже зарегистрирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                loginAlreadyExists:
                  summary: Конфликт канонического login
                  value:
                    code: login_already_exists
                    message: login is already registered
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/auth/login:
    post:
      summary: Войти в online-сессию
      description: >-
        Проверяет login и password зарегистрированного пользователя и возвращает bearer token. Login
        канонизируется по тем же правилам, что и при регистрации. Неверный login, неизвестный пользователь и
        неверный password возвращают единый ответ invalid_credentials. Password не возвращается в ответе.
        Максимальный размер тела запроса — 4 MiB.
      operationId: loginUser
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              login: alice
              password: correct-horse-battery-staple
      responses:
        '200':
          description: Пользователь аутентифицирован
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
            Pragma:
              $ref: '#/components/headers/PragmaNoCache'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              example:
                access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.example.signature
                expires_at: '2026-07-01T12:15:00Z'
                user:
                  id: 42
                  login: alice
                  created_at: '2026-07-01T12:00:00Z'
        '400':
          description: Некорректный JSON login-запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidRequest:
                  summary: Некорректный login-запрос
                  value:
                    code: invalid_request
                    message: invalid login request
        '401':
          description: Login или password не прошли проверку
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidCredentials:
                  summary: Неверные учётные данные
                  value:
                    code: invalid_credentials
                    message: invalid login or password
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/users/me:
    get:
      summary: Получить текущего пользователя
      description: >-
        Возвращает публичные данные пользователя, соответствующего bearer token из заголовка Authorization.
        Если token отсутствует, некорректен, просрочен или пользователь больше не найден, Сервер возвращает
        unauthorized.
      operationId: getCurrentUser
      tags:
        - Users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Текущий пользователь найден
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
            Pragma:
              $ref: '#/components/headers/PragmaNoCache'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              example:
                id: 42
                login: alice
                created_at: '2026-07-01T12:00:00Z'
        '401':
          description: Bearer token отсутствует или не прошёл проверку
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
            Pragma:
              $ref: '#/components/headers/PragmaNoCache'
            WWW-Authenticate:
              description: Ожидаемая Сервером схема аутентификации.
              schema:
                type: string
                example: Bearer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                unauthorized:
                  summary: Отсутствующий или некорректный token
                  value:
                    code: unauthorized
                    message: missing or invalid bearer token
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/records:
    post:
      summary: Создать приватную запись
      description: >-
        Создаёт новую приватную запись текущего пользователя. Поддерживаются type text, credentials, card и
        binary. Payload передаётся по HTTPS, валидируется Сервером, шифруется AES-GCM и сохраняется в
        PostgreSQL только в виде ciphertext. Metadata внутри payload является приватной и хранится
        зашифрованной.
      operationId: createRecord
      tags:
        - Records
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecordRequest'
            examples:
              textRecord:
                summary: Text-запись
                value:
                  type: text
                  title: Alice private note
                  payload:
                    text: secret note
                    metadata: private metadata
              credentialsRecord:
                summary: Credentials-запись
                value:
                  type: credentials
                  title: GitHub
                  payload:
                    login: alice
                    password: correct-horse-battery-staple
                    url: https://github.com/login
                    metadata: private metadata
              cardRecord:
                summary: Card-запись
                value:
                  type: card
                  title: "Joel's card"
                  payload:
                    number: '2013061420200619'
                    cardholder: Joel Miller
                    expiry_month: 3
                    expiry_year: 38
                    cvv: '014'
                    metadata: test card
              binaryRecord:
                summary: Binary-запись
                value:
                  type: binary
                  title: Alice encrypted backup
                  payload:
                    filename: backup.bin
                    data: AAEC/w==
                    metadata: private backup
      responses:
        '201':
          description: Запись создана
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
            Pragma:
              $ref: '#/components/headers/PragmaNoCache'
            ETag:
              description: Текущая revision созданной записи.
              schema:
                type: string
                example: '"1"'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordResponse'
              examples:
                textRecord:
                  $ref: '#/components/examples/TextRecordResponse'
                credentialsRecord:
                  $ref: '#/components/examples/CredentialsRecordResponse'
                cardRecord:
                  $ref: '#/components/examples/CardRecordResponse'
                binaryRecord:
                  $ref: '#/components/examples/BinaryRecordResponse'
        '400':
          description: Некорректный JSON, тип записи, title или payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidRequest:
                  summary: Некорректные данные записи
                  value:
                    code: invalid_request
                    message: invalid record request
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      summary: Получить список приватных записей
      description: Возвращает открытые metadata записей текущего пользователя без расшифрования payload'ов.
      operationId: listRecords
      tags:
        - Records
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Список metadata записей
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
            Pragma:
              $ref: '#/components/headers/PragmaNoCache'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListRecordsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/records/{id}:
    get:
      summary: Получить приватную запись
      description: >-
        Возвращает открытую metadata и расшифрованный payload одной записи текущего пользователя. Формат
        payload определяется полем type. Отсутствующая и принадлежащая другому пользователю запись возвращают
        одинаковый 404.
      operationId: getRecord
      tags:
        - Records
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID записи.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Запись найдена
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
            Pragma:
              $ref: '#/components/headers/PragmaNoCache'
            ETag:
              description: Текущая revision записи.
              schema:
                type: string
                example: '"1"'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordResponse'
              examples:
                textRecord:
                  $ref: '#/components/examples/TextRecordResponse'
                credentialsRecord:
                  $ref: '#/components/examples/CredentialsRecordResponse'
                cardRecord:
                  $ref: '#/components/examples/CardRecordResponse'
                binaryRecord:
                  $ref: '#/components/examples/BinaryRecordResponse'
        '400':
          description: Некорректный UUID записи
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidRequest:
                  summary: Некорректный record id
                  value:
                    code: invalid_request
                    message: invalid record request
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/RecordNotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      summary: Изменить приватную запись
      description: >-
        Заменяет открытый title и приватный payload существующей записи без изменения её type. Клиент обязан
        передать актуальную revision записи в заголовке If-Match как strong ETag, например "1". Сервер
        выполняет update только при совпадении ожидаемой и текущей revision, затем атомарно увеличивает
        revision и возвращает новое значение в ETag. Отсутствующая и принадлежащая другому пользователю запись
        возвращают одинаковый 404.
      operationId: updateRecord
      tags:
        - Records
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID записи.
          schema:
            type: string
            format: uuid
        - name: If-Match
          in: header
          required: true
          description: Ожидаемая revision записи в формате strong ETag, например "1".
          schema:
            type: string
            pattern: '^"[1-9][0-9]*"$'
            example: '"1"'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRecordRequest'
            examples:
              textRecord:
                summary: Изменение text-записи
                value:
                  type: text
                  title: Alice updated note
                  payload:
                    text: updated secret note
                    metadata: updated private metadata
              credentialsRecord:
                summary: Изменение credentials-записи
                value:
                  type: credentials
                  title: GitHub updated
                  payload:
                    login: alice@example.com
                    password: new-correct-horse-battery-staple
                    url: https://github.com/login
                    metadata: updated private metadata
              cardRecord:
                summary: Изменение card-записи
                value:
                  type: card
                  title: "Joel's card updated"
                  payload:
                    number: '2013061420200619'
                    cardholder: Joel Miller
                    expiry_month: 3
                    expiry_year: 38
                    cvv: '014'
                    metadata: test card updated
              binaryRecord:
                summary: Изменение binary-записи
                value:
                  type: binary
                  title: Alice updated backup
                  payload:
                    filename: backup-v2.bin
                    data: /wIBAA==
                    metadata: updated private backup
      responses:
        '200':
          description: Запись изменена
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
            Pragma:
              $ref: '#/components/headers/PragmaNoCache'
            ETag:
              description: Новая revision записи после успешного update.
              schema:
                type: string
                example: '"2"'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordResponse'
              examples:
                textRecord:
                  $ref: '#/components/examples/TextRecordResponse'
                credentialsRecord:
                  $ref: '#/components/examples/CredentialsRecordResponse'
                cardRecord:
                  $ref: '#/components/examples/CardRecordResponse'
                binaryRecord:
                  $ref: '#/components/examples/BinaryRecordResponse'
        '400':
          description: Некорректный UUID записи, If-Match, JSON, тип записи, title или payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidRequest:
                  summary: Некорректные данные изменения записи
                  value:
                    code: invalid_request
                    message: invalid record request
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/RecordNotFound'
        '409':
          $ref: '#/components/responses/RecordRevisionConflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      summary: Удалить приватную запись
      description: >-
        Физически удаляет запись текущего пользователя. Клиент обязан передать актуальную revision записи в
        заголовке If-Match как strong ETag, например "1". Сервер удаляет запись только при совпадении
        ожидаемой и текущей revision. Отсутствующая и принадлежащая другому пользователю запись возвращают
        одинаковый 404.
      operationId: deleteRecord
      tags:
        - Records
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID записи.
          schema:
            type: string
            format: uuid
        - name: If-Match
          in: header
          required: true
          description: Ожидаемая revision записи в формате strong ETag, например "1".
          schema:
            type: string
            pattern: '^"[1-9][0-9]*"$'
            example: '"1"'
      responses:
        '204':
          description: Запись физически удалена
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
            Pragma:
              $ref: '#/components/headers/PragmaNoCache'
        '400':
          description: Некорректный UUID записи или If-Match
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidRequest:
                  summary: Некорректные данные удаления записи
                  value:
                    code: invalid_request
                    message: invalid record request
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/RecordNotFound'
        '409':
          $ref: '#/components/responses/RecordRevisionConflict'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT bearer token из ответа `POST /api/v1/auth/login`.
  headers:
    CacheControlNoStore:
      description: Приватный ответ не должен сохраняться в HTTP-кэшах.
      schema:
        type: string
        example: no-store
    PragmaNoCache:
      description: HTTP/1.0-совместимый запрет кеширования приватного ответа.
      schema:
        type: string
        example: no-cache
  responses:
    Unauthorized:
      description: Отсутствующий, некорректный или просроченный bearer token
      headers:
        Cache-Control:
          $ref: '#/components/headers/CacheControlNoStore'
        Pragma:
          $ref: '#/components/headers/PragmaNoCache'
        WWW-Authenticate:
          description: Ожидаемая Сервером схема аутентификации.
          schema:
            type: string
            example: Bearer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              summary: Отсутствующий или некорректный token
              value:
                code: unauthorized
                message: missing or invalid bearer token
    RecordNotFound:
      description: Запись не найдена или недоступна текущему пользователю
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            recordNotFound:
              summary: Запись не найдена
              value:
                code: record_not_found
                message: record not found
    RecordRevisionConflict:
      description: Текущая revision записи не совпадает с revision из If-Match
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            recordRevisionConflict:
              summary: Запись была изменена другим Клиентом
              value:
                code: record_revision_conflict
                message: record revision conflict
    PreconditionRequired:
      description: Для изменения или удаления записи требуется заголовок If-Match
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            preconditionRequired:
              summary: Не передана ожидаемая revision записи
              value:
                code: precondition_required
                message: record revision is required
    PayloadTooLarge:
      description: >-
        Превышен лимит: request body — 4 MiB, text — 1 MiB или decoded binary data — 2 MiB.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            payloadTooLarge:
              summary: Превышен допустимый размер запроса или payload
              value:
                code: payload_too_large
                message: payload is too large
    UnsupportedMediaType:
      description: Запрос передан не в формате JSON
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unsupportedMediaType:
              summary: Неподдерживаемый Content-Type
              value:
                code: unsupported_media_type
                message: content type must be application/json
    InternalError:
      description: Внутренняя ошибка Сервера
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            internalError:
              summary: Внутренняя ошибка без раскрытия деталей
              value:
                code: internal_error
                message: internal server error
  examples:
    TextRecordResponse:
      summary: Text-запись
      value:
        id: 550e8400-e29b-41d4-a716-446655440000
        type: text
        title: Alice private note
        revision: 1
        created_at: '2026-07-08T12:00:00Z'
        updated_at: '2026-07-08T12:00:00Z'
        payload:
          text: secret note
          metadata: private metadata
    CredentialsRecordResponse:
      summary: Credentials-запись
      value:
        id: 550e8400-e29b-41d4-a716-446655440001
        type: credentials
        title: GitHub
        revision: 1
        created_at: '2026-07-10T12:00:00Z'
        updated_at: '2026-07-10T12:00:00Z'
        payload:
          login: alice
          password: correct-horse-battery-staple
          url: https://github.com/login
          metadata: private metadata
    CardRecordResponse:
      summary: Card-запись
      value:
        id: 550e8400-e29b-41d4-a716-446655440002
        type: card
        title: "Joel's card"
        revision: 1
        created_at: '2026-07-11T12:00:00Z'
        updated_at: '2026-07-11T12:00:00Z'
        payload:
          number: '2013061420200619'
          cardholder: Joel Miller
          expiry_month: 3
          expiry_year: 38
          cvv: '014'
          metadata: test card
    BinaryRecordResponse:
      summary: Binary-запись
      value:
        id: 550e8400-e29b-41d4-a716-446655440003
        type: binary
        title: Alice encrypted backup
        revision: 1
        created_at: '2026-07-12T12:00:00Z'
        updated_at: '2026-07-12T12:00:00Z'
        payload:
          filename: backup.bin
          data: AAEC/w==
          metadata: private backup
  schemas:
...
    RegisterRequest:
      type: object
      additionalProperties: false
      required:
        - login
        - password
      properties:
        login:
          $ref: '#/components/schemas/LoginField'
        password:
          $ref: '#/components/schemas/RegistrationPasswordField'
    LoginRequest:
      type: object
      additionalProperties: false
      required:
        - login
        - password
      properties:
        login:
          $ref: '#/components/schemas/LoginField'
        password:
          $ref: '#/components/schemas/LoginPasswordField'
    LoginResponse:
      type: object
      additionalProperties: false
      required:
        - access_token
        - expires_at
        - user
      properties:
        access_token:
          type: string
          description: JWT access token для заголовка Authorization со схемой Bearer.
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.example.signature
        expires_at:
          type: string
          format: date-time
          description: Момент истечения access token в UTC.
          example: '2026-07-01T12:15:00Z'
        user:
          $ref: '#/components/schemas/UserResponse'
    UserResponse:
      type: object
      additionalProperties: false
      required:
        - id
        - login
        - created_at
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
          example: 42
        login:
          $ref: '#/components/schemas/CanonicalLogin'
        created_at:
          type: string
          format: date-time
          example: '2026-07-01T12:00:00Z'
    LoginField:
      type: string
      description: >-
        Login пользователя во входном запросе. Сервер сначала удаляет пробельные символы по краям и переводит
        ASCII-буквы A-Z в нижний регистр. Ограничения длины и допустимых символов применяются к полученному
        каноническому значению, описанному схемой CanonicalLogin.
      example: '  Alice  '
    CanonicalLogin:
      type: string
      description: >-
        Канонический login пользователя после удаления пробельных символов по краям и преобразования ASCII-букв
        A-Z в нижний регистр. Содержит от 3 до 32 ASCII-символов: буквы, цифры, точку, подчёркивание или дефис;
        первый символ — буква или цифра.
      minLength: 3
      maxLength: 32
      pattern: '^[a-z0-9][a-z0-9._-]{2,31}$'
      example: alice
    RegistrationPasswordField:
      type: string
      format: password
      writeOnly: true
      minLength: 8
      maxLength: 64
      pattern: '^[!-~]{8,64}$'
      description: >-
        Password нового пользователя длиной от 8 до 64 символов. Разрешены печатные ASCII-символы от ! до ~;
        пробелы и другие Unicode-символы запрещены. Политика применяется при регистрации.
      example: correct-horse-battery-staple
    LoginPasswordField:
      type: string
      format: password
      writeOnly: true
      description: >-
        Password зарегистрированного пользователя. Сервер не применяет текущую регистрационную политику длины
        повторно, поэтому ранее созданные учётные записи с более коротким password продолжают работать.
      example: correct-horse-battery-staple
...
    CreateRecordRequest:
      oneOf:
        - $ref: '#/components/schemas/CreateTextRecordRequest'
        - $ref: '#/components/schemas/CreateCredentialsRecordRequest'
        - $ref: '#/components/schemas/CreateCardRecordRequest'
        - $ref: '#/components/schemas/CreateBinaryRecordRequest'
      discriminator:
        propertyName: type
        mapping:
          text: '#/components/schemas/CreateTextRecordRequest'
          credentials: '#/components/schemas/CreateCredentialsRecordRequest'
          card: '#/components/schemas/CreateCardRecordRequest'
          binary: '#/components/schemas/CreateBinaryRecordRequest'
...
    UpdateRecordRequest:
      oneOf:
        - $ref: '#/components/schemas/UpdateTextRecordRequest'
        - $ref: '#/components/schemas/UpdateCredentialsRecordRequest'
        - $ref: '#/components/schemas/UpdateCardRecordRequest'
        - $ref: '#/components/schemas/UpdateBinaryRecordRequest'
      discriminator:
        propertyName: type
        mapping:
          text: '#/components/schemas/UpdateTextRecordRequest'
          credentials: '#/components/schemas/UpdateCredentialsRecordRequest'
          card: '#/components/schemas/UpdateCardRecordRequest'
          binary: '#/components/schemas/UpdateBinaryRecordRequest'
...
    RecordResponse:
      oneOf:
        - $ref: '#/components/schemas/TextRecordResponse'
        - $ref: '#/components/schemas/CredentialsRecordResponse'
        - $ref: '#/components/schemas/CardRecordResponse'
        - $ref: '#/components/schemas/BinaryRecordResponse'
      discriminator:
        propertyName: type
        mapping:
          text: '#/components/schemas/TextRecordResponse'
          credentials: '#/components/schemas/CredentialsRecordResponse'
          card: '#/components/schemas/CardRecordResponse'
          binary: '#/components/schemas/BinaryRecordResponse'
...
    ListRecordsResponse:
      type: object
      additionalProperties: false
      required:
        - records
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/RecordMetadataResponse'
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Стабильный машиночитаемый код ошибки.
          example: invalid_request
        message:
          type: string
          description: Безопасное сообщение без внутренних деталей и секретов.
          example: invalid registration data